Big update to parser. Allows it to be rewindable.
Should have committed more frequently but, too late now.

Creating a new "rewindable" contextmanager on the Token class.
This allows us to try an operation to see if it resolves and if it doesn't, we
don't have to worry about having consumed part of the stream of tokens.

I am also introducing the concept of an exception stack on the Token class
object.
This allows us to store an exception on there for inspection later when we
capture it from the rewindable contextmanager.
We can now query why there are still unparsed tokens left at the end of a parse
call.

We have created a few of our own exception types for general good practice.
We have removed the OrToken specific reimplementation of resolve because it is
now functionally the same as the Token one with our rewindable contextmanager.

Our parser is now ignoring any tokens that the lexer sends it with the tokenType
of "IGNORE". This now includes newlines as well as whitespace.
TODO: Make sure our string token is higher precedence than the IGNORE tokens.

We also throw a ParsingError at the end of a parse if we did not consume all of
the tokens.