Updating the parser to support datatypes and the bnf parser for literal values.
Creating a "LiteralToken" which is a new feature for the bnf parser. If you separate a token in literal quotes, it will be interpreted as a literal value.
Also updating most, if not all, of the indexing done against the old LexerToken tuple to use the new names from the namedtuple.
Adding a "cast" concept to the values in knobs so that I can cast them into the correct datatype for working with.async